+++ src/components/Match/CastTable.jsx
@@ -11,6 +11,22 @@ import {
 } from 'config';
 import Table from '../Table/Table';
 
+// TODO localize strings
+
+const castsColumns = [{
+  displayName: 'Name',
+  field: 'name',
+}, {
+  displayName: 'Casts',
+  field: 'casts',
+}, {
+  displayName: 'Hits',
+  field: 'hero_hits',
+}, {
+  displayName: 'Damage',
+  field: 'damage_inflictor',
+}];
+
 const getCastArray = (pm) => {
   // Get from ability_uses, item_uses
   const resultArray = [];
@@ -34,7 +50,6 @@ const getCastArray = (pm) => {
 
 const CastTable = ({
   match,
-  columns,
 }) => (
   <Tabs>
     {match.players.map(p =>
@@ -42,7 +57,7 @@ const CastTable = ({
       <Tab key={p.player_slot} icon={<img src={`${API_HOST}${heroes[p.hero_id].img}`} height={30} role="presentation" />}>
         <Table
           data={getCastArray(p)}
-          columns={columns}
+          columns={castsColumns}
         />
       </Tab>
       ))
