+++ src/components/Player/Pages/Matches/playerMatchesColumns.jsx
@@ -2,54 +2,54 @@ import { transformations } from 'utility';
 import strings from 'lang';
 
 export default [{
-  displayName: strings.th_hero,
-  tooltip: strings.hero_id,
+  displayName: strings.th_hero_id,
+  tooltip: strings.tooltip_hero_id,
   field: 'hero_id',
   width: 2.8,
   displayFn: transformations.hero_id,
 }, {
-  displayName: strings.th_id,
-  tooltip: strings.match_id,
+  displayName: strings.th_match_id,
+  tooltip: strings.tooltip_match_id,
   field: 'match_id',
   width: 1.8,
   sortFn: true,
   displayFn: transformations.match_id_and_game_mode,
 }, {
   displayName: strings.th_result,
-  tooltip: strings.result,
+  tooltip: strings.tooltip_result,
   field: 'radiant_win',
   width: 1.5,
   displayFn: transformations.radiant_win,
 }, {
   displayName: strings.th_skill,
-  tooltip: strings.skill,
+  tooltip: strings.tooltip_skill,
   field: 'skill',
   width: 1.5,
   sortFn: true,
   displayFn: transformations.skill,
 }, {
   displayName: strings.th_duration,
-  tooltip: strings.duration,
+  tooltip: strings.tooltip_duration,
   field: 'duration',
   width: 1.2,
   sortFn: true,
   displayFn: transformations.duration,
 }, {
   displayName: strings.abbr_kills,
-  tooltip: strings.kills,
+  tooltip: strings.tooltip_kills,
   field: 'kills',
   sortFn: true,
   width: 0.8,
   displayFn: transformations.kda,
 }, {
   displayName: strings.abbr_deaths,
-  tooltip: strings.deaths,
+  tooltip: strings.tooltip_deaths,
   field: 'deaths',
   sortFn: true,
   width: 0.8,
 }, {
   displayName: strings.abbr_assists,
-  tooltip: strings.assists,
+  tooltip: strings.tooltip_assists,
   field: 'assists',
   sortFn: true,
   width: 0.8,
