+++ src/components/Player/Player.jsx
@@ -9,73 +9,11 @@ import {
   getPlayer,
   getPlayerWinLoss,
 } from 'actions';
-import strings from 'lang';
 import TabBar from 'components/TabBar';
 import PlayerHeader from './PlayerHeader';
 import Error from '../Error';
 import styles from './Player.css';
-import {
-  PeersPage,
-  OverviewPage,
-  MatchesPage,
-  HeroesPage,
-  ProsPage,
-  RankingsPage,
-  HistogramsPage,
-  RecordsPage,
-  CountsPage,
-  TrendsPage,
-  MMRPage,
-} from './Pages';
-
-const playerPages = [{
-  name: strings.tab_overview,
-  content: playerId => (<OverviewPage playerId={playerId} />),
-}, {
-  name: strings.tab_matches,
-  content: playerId => (<MatchesPage playerId={playerId} />),
-}, {
-  name: strings.tab_heroes,
-  content: playerId => (<HeroesPage playerId={playerId} />),
-}, {
-  name: strings.tab_peers,
-  content: playerId => (<PeersPage playerId={playerId} />),
-}, {
-  name: strings.tab_pros,
-  content: playerId => (<ProsPage playerId={playerId} />),
-}, {
-  name: strings.tab_activity,
-  content: () => (<div />),
-}, {
-  name: strings.tab_records,
-  content: playerId => (<RecordsPage playerId={playerId} />),
-}, {
-  name: strings.tab_counts,
-  content: playerId => (<CountsPage playerId={playerId} />),
-}, {
-  name: strings.tab_histograms,
-  content: (playerId, subInfo) => (<HistogramsPage playerId={playerId} histogramName={subInfo} />),
-}, {
-  name: strings.tab_trends,
-  content: (playerId, subInfo) => (<TrendsPage playerId={playerId} trendName={subInfo} />),
-}, {
-  name: strings.tab_wardmap,
-  content: () => (<div />),
-}, {
-  name: strings.tab_wordcloud,
-  content: () => (<div />),
-}, {
-  name: strings.tab_mmr,
-  content: playerId => (<MMRPage playerId={playerId} />),
-}, {
-  name: strings.tab_rankings,
-  content: playerId => (<RankingsPage playerId={playerId} />),
-}];
-
-const playerPagesMapped = playerId => playerPages.map(page => ({
-  ...page,
-  route: `/players/${playerId}/${page.name.toLowerCase()}`,
-}));
+import playerPages from './playerPages';
 
 const Player = ({
   params: {
@@ -88,7 +26,7 @@ const Player = ({
     return <Error />;
   }
   // Need to pass in the action into filter form, need to put that filter form into each subroute as well
-  const page = playerPagesMapped(accountId).find(page => page.name.toLowerCase() === info);
+  const page = playerPages(accountId).find(page => page.name.toLowerCase() === info);
   return (
     <div>
       <div className={styles.header}>
@@ -97,7 +35,7 @@ const Player = ({
           <TabBar
             info={info}
             subInfo={subInfo}
-            tabs={playerPagesMapped(accountId)}
+            tabs={playerPages(accountId)}
           />
         </div>
       </div>
